add unit tests for all newly created mappers

+++ src/main/java/at/qe/skeleton/dtos/ClimateDataPointDTO.java
@@ -6,6 +6,6 @@ public record ClimateDataPointDTO(
         LocalDateTime timestamp,
         double temperature,
         double humidity,
-        int airQuality
+        double airQuality
 ) {
 }
